+++ src/domain/accessibilityStatement/AccessibilityStatementEn.tsx
@@ -5,13 +5,14 @@ const AccessibilityStatementEn: FunctionComponent = () => {
     <>
       <h1>Accessibility Statement</h1>
       <p>
-        This accessibility statement applies to the website Culture Kids of the
-        City of Helsinki. The site address is https://kummilapset.hel.fi
+        This accessibility statement applies to the website ‘Culture Kids‘ of
+        the City of Helsinki. The site address is http://kummilapset.hel.fi/.
       </p>
 
       <h2>Statutory provisions applicable to the website</h2>
       <p>
-        This website was published after 23 September 2018. The website must
+        This website was published after{' '}
+        <time dateTime="2018-09-23">23 September 2018</time>. The website must
         fulfil statutory accessibility requirements.
       </p>
 
@@ -24,17 +25,88 @@ const AccessibilityStatementEn: FunctionComponent = () => {
 
       <h2>Compliance status</h2>
       <p>
-        The accessibility compliance of this website has not yet been evaluated.
+        This website fulfils the critical legal accessibility requirements in
+        accordance with level AA of WCAG v2.1. The shortcomings observed in the
+        audit have been corrected as suggested.
       </p>
 
-      <h2>Preparing an accessibility statement</h2>
-      <p>This statement was prepared on 17.1.2020</p>
+      <h2>Correction shortcomings</h2>
+      <p>
+        The shortcomings observed in the audit have been corrected according to
+        the suggestions. The website has not been re-audited since the
+        corrections were made. If the service still contains inaccessible
+        content, please contact the parties mentioned above. We will be taking
+        into account the factors mentioned above in the further development of
+        the service in connection with each new publication.
+      </p>
+
+      <h2>Availability of information in an accessible format</h2>
+      <p>
+        Access to any content that cannot be accessed due to the shortcomings
+        mentioned earlier can be requested from the administrator of the
+        website.
+      </p>
+      <p>Culture and Leisure Division</p>
+
+      <h2>Preparation of this accessibility statement</h2>
+      <p>
+        This statement was prepared on{' '}
+        <time dateTime="2020-03-23">23 March 2020</time> and updated on{' '}
+        <time dateTime="2020-09-09">9 September 2020</time>.
+      </p>
+
+      <h3>Assessment of accessibility</h3>
+      <p>
+        The working instruction and procedures of the City of Helsinki were
+        followed when evaluating the accessibility of the site, with the aim of
+        ensuring that websites are accessible in all stages of the work process.
+      </p>
+      <p>
+        Accessibility was evaluated by means of an audit by a third-party expert
+        as well as self-evaluation.
+      </p>
+      <p>
+        Accessibility was evaluated using a programmatic accessibility auditing
+        tool as well as by manually reviewing the site and content. The
+        programmatic assessment was carried out by using an automatic
+        accessibility testing tool and a browser extension provided by
+        Siteimprove.
+      </p>
+      <p>
+        The third party expert audit was carried out by the City of Helsinki
+        Service Centre enterprise.
+      </p>
 
       <h3>Updating the accessibility statement</h3>
       <p>
-        The accessibility statement will be updated so that it corresponds with
-        the observations related to accessibility compliance made during an
-        audit.
+        When website technology or content changes, its accessibility must be
+        ensured through constant monitoring and periodic checks. This statement
+        will be updated in conjunction with website changes and accessibility
+        evaluations.
+      </p>
+
+      <h2>Feedback and contact information</h2>
+      <p>
+        The party responsible for the accessibility of the website is the
+        Communications and Marketing Service of the City of Helsinki’s Culture
+        and Leisure Division, telephone exchange 09 310 1060.
+      </p>
+      <p>
+        City of Helsinki, Culture and Leisure Division
+        <br /> Culture Kids
+        <br />
+        kulttuurin.kummilapset@hel.fi
+      </p>
+
+      <h3>Reporting non-accessible content</h3>
+      <p>
+        If a user feels that accessibility requirements have not been met, they
+        can report the issue by e-mail to{' '}
+        <a href="mailto:helsinki.palaute@hel.fi">helsinki.palaute@hel.fi</a> or
+        through the feedback form at{' '}
+        <a href="https://www.hel.fi/helsinki/en/administration/participate/feedback">
+          www.hel.fi/palaute
+        </a>
       </p>
 
       <h3>Requesting information in an accessible format</h3>
@@ -43,18 +115,36 @@ const AccessibilityStatementEn: FunctionComponent = () => {
         accessible format, they can request for this information by e-mail at{' '}
         <a href="mailto:helsinki.palaute@hel.fi">helsinki.palaute@hel.fi</a> or
         through the{' '}
-        <a href="https://hel.fi/helsinki/en/administration/participate/feedback">
+        <a href="https://www.hel.fi/helsinki/en/administration/participate/feedback">
           feedback form
         </a>
         . The aim is to reply to the enquiry within a reasonable time frame.
       </p>
 
-      <h2>Feedback and contact information</h2>
+      <h2>Legal protection of accessibility, Enforcement procedure</h2>
       <p>
-        City of Helsinki, Culture and Leisure Division
-        <br /> Culture Kids
+        If a user feels that their report or enquiry has not received a response
+        or that the response is unsatisfactory, they can report the issue to the
+        Regional State Administrative Agency of Southern Finland. The website of
+        the Regional State Administrative Agency of Southern Finland explains in
+        detail how the matter will be processed.
+      </p>
+
+      <p>
+        Regional State Administrative Agency of Southern Finland
         <br />
-        kulttuurin.kummilapset@hel.fi
+        Accessibility monitoring unit
+        <br />
+        <a href="https://www.saavutettavuusvaatimukset.fi">
+          www.saavutettavuusvaatimukset.fi
+        </a>{' '}
+        (in Finnish)
+        <br />
+        <a href="mailto:saavutettavuus@avi.fi">saavutettavuus@avi.fi </a>
+        <br />
+        Telephone exchange <a href="tel:+358295016000">+358 295 016 000</a>
+        <br />
+        Open: Mon-Fri at 8:00–16:15
       </p>
 
       <h2>The City of Helsinki and accessibility</h2>
@@ -81,8 +171,24 @@ const AccessibilityStatementEn: FunctionComponent = () => {
         assistive technologies. Support is available on guidance sites announced
         on the city’s website and through telephone counselling.
       </p>
+
+      <p>
+        <strong>Information on City services</strong>
+        <br />
+        Telephone guidance: <a href="tel:+358931011111">+358 (0)9 310 11111</a>
+        <br />
+        Opening hours: Mon–Thu 9–16 and Fri 10–15
+        <br />
+        <a href="https://www.hel.fi/helsinki/en/administration/administration/communication">
+          Go to the City’s guidance site
+        </a>
+      </p>
+
       <h2>Approval of the accessibility statement</h2>
-      <p>This statement was approved on 17.1.2020</p>
+      <p>
+        This statement was approved on{' '}
+        <time dateTime="2020-09-09">9 September 2020</time> by
+      </p>
       <p>
         Culture and Leisure Division <br />
         City of Helsinki.